Cleanup & add compiler warning options.
- Add compiler warning. This avoids common problems.

- Add a top-level CMakeLists. This allows developers to know they can
  use CMake and can directly use cmake using `cmake ..` from the build/
  directory, instead of cmake `../board`.

- Merge the few Makefile into a shorter one.

- Remove the bin directory. Output directly into the build directory.

- Add tools/format.sh to run clang-format on every files.

- Format every files.

+++ src/ui/ui.cpp
@@ -70,7 +70,7 @@ class MainMenu : public ComponentBase {
     menu_.resize(menu_group.size());
     tab_.resize(menu_group.size());
 
-    for (int i = 0; i < menu_group.size(); ++i) {
+    for (size_t i = 0; i < menu_group.size(); ++i) {
       menu_[i] = Menu(&menu_entries_[i], &index_[i]);
       tab_[i] = Container::Tab({}, &index_[i]);
       index_[i] = 0;
